[sosreport] remove tmpdir only if it exists
Prevents raising exception when the dir has been removed.

+++ sos/sosreport.py
@@ -1611,7 +1611,7 @@ class SoSReport(object):
         logging.shutdown()
         if self.tempfile_util:
             self.tempfile_util.clean()
-        if self.tmpdir:
+        if self.tmpdir and os.path.isdir(self.tmpdir):
             rmtree(self.tmpdir)
 
         return True
